+++ internal/sonos/services.go
@@ -4,9 +4,11 @@ const (
 	controlAVTransport       = "/MediaRenderer/AVTransport/Control"
 	controlRenderingControl  = "/MediaRenderer/RenderingControl/Control"
 	controlGroupRendering    = "/MediaRenderer/GroupRenderingControl/Control"
+	controlContentDirectory  = "/MediaServer/ContentDirectory/Control"
 	controlZoneGroupTopology = "/ZoneGroupTopology/Control"
 	urnAVTransport           = "urn:schemas-upnp-org:service:AVTransport:1"
 	urnRenderingControl      = "urn:schemas-upnp-org:service:RenderingControl:1"
 	urnGroupRenderingControl = "urn:schemas-upnp-org:service:GroupRenderingControl:1"
+	urnContentDirectory      = "urn:schemas-upnp-org:service:ContentDirectory:1"
 	urnZoneGroupTopology     = "urn:schemas-upnp-org:service:ZoneGroupTopology:1"
 )